requirements.txt
# Models
#pytorch # Install manually!

# Necessary
matplotlib
tqdm
shapely
squarify
torchnet
watermark
attrdict
scikit-learn
Pillow # For displaying images in notebooks
colored # Display

# Optional NLP tools (needed for NLP task)
nltk
torchtext
spacy
seaborn
